advanced-python-analytics/requirements.txt
# Core Data Science
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=1.7.0

# Visualization
plotly>=5.14.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Database
sqlalchemy>=2.0.0

# Async
aiohttp>=3.8.0
aiofiles>=23.1.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Code Quality
black>=23.3.0
flake8>=6.0.0
mypy>=1.4.0
pylint>=2.17.0

# Utilities
python-dateutil>=2.8.0
tqdm>=4.65.0
joblib>=1.3.0

# Jupyter (optional for notebooks)
jupyter>=1.0.0
ipykernel>=6.23.0
